feat: add Link 'Log in' and 'Sign in'
