--- src/components/header/Header.stories.ts
@@ -1,25 +1,0 @@
-import { Meta, StoryObj } from '@storybook/react'
-
-import { Header } from './Header'
-
-const meta = {
-  component: Header,
-  tags: ['autodocs'],
-  title: 'Components/Header',
-} satisfies Meta<typeof Header>
-
-export default meta
-type Story = StoryObj<typeof meta>
-
-export const Default: Story = {
-  args: {
-    countNotification: 3,
-    isAuth: true,
-  },
-}
-
-export const NotAutorealized: Story = {
-  args: {
-    isAuth: false,
-  },
-}
