+++ src/components/select/Select.module.scss
@@ -1,6 +1,5 @@
 .container {
-  // width: 100%;
-  display: flex; //добавил
+  // width: 100%; тут он тоже мне мешал
 }
 
 .label {
